Avances del proyecto, clases creadas(servicios, repositorios,...) y avances en cuanto a los equipos

+++ Proyectos/MotoGPPerezMarquezGabriel/src/main/java/com/salesianostriana/dam/motogpperezmarquezgabriel/model/Mecanico.java
@@ -16,6 +16,6 @@ public class Mecanico {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	private String nombre;
-	private float cuota;
+	private double cuota;
 	
 }
